Update: created utils.py, moved get_days_on_vacation and create_token to it
