+++ app/controllers/login.py
@@ -1,7 +1,6 @@
 from flask import jsonify, request
 from app.database import get_user_by_email_and_password
-import jwt
-import datetime
+from utils import create_token
 
 def employee_login():
     data = request.json
@@ -15,12 +14,3 @@ def employee_login():
 
     token = create_token(user, "employee")
     return jsonify({"token": token})
-
-def create_token(user, role):
-    payload = {
-        "email": user.user_email,
-        "role": role,
-        "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=2)
-    }
-    token = jwt.encode(payload, "hardcoded", algorithm="HS256")
-    return token
