+++ app/controllers/upload_files.py
@@ -1,7 +1,7 @@
 from flask import jsonify, request
 from io import StringIO
 from app.database import create_user, create_used_vacations, create_available_vacations
-from datetime import datetime
+from utils import get_days_on_vacation
 import csv
 
 def upload_users_file():
@@ -56,18 +56,6 @@ def upload_used_vacation_file():
         )
 
     return {"Vacations added: ": len(used_vacations)}, 201
-
-
-# Move this to utils file
-def get_days_on_vacation(start_date: str, end_date: str):
-    start = datetime.strptime(start_date, "%A, %B %d, %Y")
-    end = datetime.strptime(end_date, "%A, %B %d, %Y")
-    delta = end - start
-    year = end.year
-    return {
-        "days_on_vacation": delta.days + 1,
-        "year": year,
-    }
 
 def upload_vacation_days_file():
     if "file" not in request.files:
