+++ app/controllers/utils.py
@@ -1,0 +1,21 @@
+import jwt, datetime
+
+def get_days_on_vacation(start_date: str, end_date: str):
+    start = datetime.strptime(start_date, "%A, %B %d, %Y")
+    end = datetime.strptime(end_date, "%A, %B %d, %Y")
+    delta = end - start
+    year = end.year
+    return {
+        "days_on_vacation": delta.days + 1,
+        "year": year,
+    }
+
+def create_token(user, role):
+    payload = {
+        "email": user.user_email,
+        "role": role,
+        "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=2)
+    }
+    token = jwt.encode(payload, "hardcoded", algorithm="HS256")
+    return token
+
